Fix new clippy warnings

+++ crates/npc-engine/src/db/props.rs
@@ -1,3 +1,22 @@
+/*
+ * niepce - engine/db/props.rs
+ *
+ * Copyright (C) 2021-2023 Hubert Figuière
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 use lazy_static::lazy_static;
 use maplit::hashmap;
 use npc_fwk::utils::exempi::NS_AUX as NS_EXIF_AUX;
@@ -19,7 +38,7 @@ pub enum NiepceProperties {
 impl From<NiepceProperties> for u32 {
     fn from(v: NiepceProperties) -> u32 {
         match v {
-            NiepceProperties::Index(i) => i.repr as u32,
+            NiepceProperties::Index(i) => i.repr,
             NiepceProperties::Other(i) => i,
         }
     }
